renamed types to AMR_types.h
