+++ src/Inciter/AMR/node_store.h
@@ -4,7 +4,7 @@
 #include <cmath>
 
 #include "Base/Exception.h"
-#include "types.h"
+#include "AMR_types.h"
 #include "tet_store.h"
 
 // TODO: make this have a base class to support multiple generator schemes
