+++ src/Inciter/AMR/tet_store.h
@@ -6,7 +6,7 @@
 
 #include "Base/Exception.h"
 
-#include "types.h"
+#include "AMR_types.h"
 #include "active_element_store.h"
 #include "master_element_store.h"
 #include "edge_store.h"
@@ -639,7 +639,6 @@ namespace AMR {
                 }
             }
 
-            /*
             void unset_marked_children(size_t parent_id)
             {
                 Refinement_State& parent = data(parent_id);
@@ -647,7 +646,7 @@ namespace AMR {
                 {
                     marked_refinements.erase(c);
                 }
-            }*/
+            }
 
             child_id_list_t generate_child_ids(size_t parent_id, size_t count = MAX_CHILDREN)
             {
